move more functions to models
