+++ src/html5sortable.ts
@@ -7,6 +7,9 @@ import { addEventListener as _on , removeEventListener as _off } from './eventLi
 import { addAttribute as _attr , removeAttribute as _removeAttr } from './attribute'
 import _offset from './offset'
 import _debounce from './debounce'
+import _index from './index'
+import _detach from './removeElement'
+import {makeElement as _html2element, insertBefore as _before, insertAfter as _after} from './insertHtmlElements'
 /*
  * variables global to the plugin
  */
@@ -243,17 +246,6 @@ var _reloadSortable = function (sortableElement) {
   _removeSortableEvents(sortableElement)
 }
 /**
- * Get position of the element relatively to its sibling elements
- * @param {Element} element
- * @returns {number}
- */
-var _index = function (element) {
-  if (!element.parentElement) {
-    return 0
-  }
-  return Array.prototype.indexOf.call(element.parentElement.children, element)
-}
-/**
  * Whether element is in DOM
  * @param {Element} element
  * @returns {boolean}
@@ -261,51 +253,6 @@ var _index = function (element) {
 var _attached = function (element) {
   // document.body.contains(element)
   return !!element.parentNode
-}
-/**
- * Convert HTML string into DOM element.
- * @param {Element|string} html
- * @param {string} tagname
- * @returns {Element}
- */
-var _html2element = function (html, tagName) {
-  if (typeof html !== 'string') {
-    return html
-  }
-  var parentElement = document.createElement(tagName)
-  parentElement.innerHTML = html
-  return parentElement.firstChild
-}
-/**
- * Insert before target
- * @param {Element} target
- * @param {Element} element
- */
-var _before = function (target, element) {
-  target.parentElement.insertBefore(
-    element,
-    target
-  )
-}
-/**
- * Insert after target
- * @param {Element} target
- * @param {Element} element
- */
-var _after = function (target, element) {
-  target.parentElement.insertBefore(
-    element,
-    target.nextElementSibling
-  )
-}
-/**
- * Detach element from DOM
- * @param {Element} element
- */
-var _detach = function (element) {
-  if (element.parentNode) {
-    element.parentNode.removeChild(element)
-  }
 }
 /**
  * Make native event that can be dispatched afterwards
@@ -529,7 +476,7 @@ export default function sortable (sortableElements, options) {
       _after(visiblePlaceholder, dragging)
     })
 
-    var debouncedDragOverEnter = _debounce(function (element, pageY) {
+    var debouncedDragOverEnter = _debounce((element, pageY) => {
       if (!dragging) {
         return
       }
@@ -574,7 +521,8 @@ export default function sortable (sortableElements, options) {
         placeholders
           .filter(function (element) { return element !== placeholder })
           .forEach(_detach)
-      } else {
+      }
+      else {
         if (placeholders.indexOf(element) === -1 &&
             !_filter(_getChildren(element), options.items).length) {
           placeholders.forEach(_detach)
@@ -582,7 +530,6 @@ export default function sortable (sortableElements, options) {
         }
       }
     }, options.debounce)
-
     // Handle dragover and dragenter events on draggable items
     var onDragOverEnter = function (e) {
       if (!dragging || !_listsConnected(sortableElement, dragging.parentElement) || _data(sortableElement, '_disabled') === 'true') {
